lombok

+++ src/main/java/com/example/testtask/service/ContactService.java
@@ -2,20 +2,16 @@ package com.example.testtask.service;
 
 import com.example.testtask.model.Contact;
 import com.example.testtask.repository.ContactRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
+import lombok.RequiredArgsConstructor;
+import lombok.var;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class ContactService {
     private final ContactRepository repo;
-
-    @Autowired
-    public ContactService(@Qualifier("contact") ContactRepository repo) {
-        this.repo = repo;
-    }
 
     public void addContact(Contact contact) {
         this.repo.save(contact);
@@ -30,11 +26,13 @@ public class ContactService {
     }
 
     public void deleteContactById(Long id) {
-        this.repo.deleteById(id);
+        if (this.repo.findById(id).isPresent()) {
+            this.repo.deleteById(id);
+        }
     }
 
     public void updateContactById(Long id, Contact updated) {
-        Optional<Contact> optionalContact = this.repo.findById(id);
+        var optionalContact = this.repo.findById(id);
         if (optionalContact.isPresent()) {
             Contact contact = optionalContact.get();
             contact.setFullName(updated.getFullName());
